# lib/api-zod/src/generated/types/createFunnelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFunnelBodyConversionGoal } from "./createFunnelBodyConversionGoal";
import type { CreateFunnelBodyTrafficSource } from "./createFunnelBodyTrafficSource";

export interface CreateFunnelBody {
  /** @nullable */
  projectId?: number | null;
  targetAudience: string;
  coreOffer: string;
  conversionGoal: CreateFunnelBodyConversionGoal;
  trafficSource: CreateFunnelBodyTrafficSource;
}