# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
}