# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProjects: number;
  draftProjects: number;
  publishedProjects: number;
  totalFunnels: number;
  averageBuildProgress: number;
  aiEngineReady: boolean;
}