# lib/api-zod/src/generated/types/funnel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FunnelAiEngineStatus } from "./funnelAiEngineStatus";
import type { FunnelConversionGoal } from "./funnelConversionGoal";
import type { FunnelTrafficSource } from "./funnelTrafficSource";

export interface Funnel {
  id: number;
  /** @nullable */
  projectId?: number | null;
  targetAudience: string;
  coreOffer: string;
  conversionGoal: FunnelConversionGoal;
  trafficSource: FunnelTrafficSource;
  buildProgress: number;
  aiEngineStatus: FunnelAiEngineStatus;
  createdAt: Date;
  updatedAt: Date;
}