# lib/api-zod/src/generated/types/updateFunnelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateFunnelBodyConversionGoal } from "./updateFunnelBodyConversionGoal";
import type { UpdateFunnelBodyTrafficSource } from "./updateFunnelBodyTrafficSource";

export interface UpdateFunnelBody {
  targetAudience?: string;
  coreOffer?: string;
  conversionGoal?: UpdateFunnelBodyConversionGoal;
  trafficSource?: UpdateFunnelBodyTrafficSource;
}