# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  name?: string;
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  description?: string | null;
}